refactor: change user interface to model
