+++ libs/data-access/auth/index.ts
@@ -1,3 +1,4 @@
 export * from './auth.module';
 export * from './services/auth.service';
 export * from './guards/auth.guard';
+export * from './models/auth.model';
