+++ libs/data-access/auth/models/auth.model.ts
@@ -1,0 +1,10 @@
+/**
+ * User interface from Firebase data authentication.
+ */
+export interface User {
+  uid: string;
+  email: string;
+  displayName: string;
+  photoURL: string;
+  emailVerified: boolean;
+}
